bin/html: detect i18n files
* bin/html (slug): Add regex to identify ISO 639-1 string in file
name. Return a tuple containing the file 'slug' and the IS0 639-1
string. The second element of the tuple is None when ISO 639-1 string
is not found in the file name.
(html): Read ISO 639-1 string into `l` and return it as the third
element.
(process): Add handling to write localized version at `SLUG/XX` where
`XX` is a ISO 639-1 string.